Makefile: Add golangci-lint, go mod tidy, and dirtyness check
The golangci-lint lint aggregator is added with a set of reasonably
opinionated linters that required a number of limited-impact changes.

A new "check" make goal runs both tests and the static checks
performed by golangci-lint. Also, new goals for running "go mod tidy"
and verifying that no dirty files are present in the workspace are
added. The point of the latter is to make sure that we're not leaving
cruft in go.{mod,sum} and that any generated file under source control
is up to date.