Build binary with make in the Docker build; build image in GH workflow
Since we have various code generation stuff going on in the makefile
we're better off using make to building during the Docker image build.
This required installing a few additional packages.

To make sure we don't accidentally break the Docker image build again
we add an image build step to the workflow that runs for commits and
pull requests.