Upgrade golangci-lint to 1.53.3 and adjust the linters used
Some linters had been deprecated and/or replaced so they had to be
removed, but we also add a bunch of new linters. Some of them found
new violations that we had to address.

The most visible change is the reformatting of the copyright notice,
required (I guess) when the golangci-lint release binary started being
built with 1.19 or whenever the formatting of indented blocks changed.

There are definitely more linters we could enable, but they would
require more significant changes to the code.