Change GitHub workflow to push built Docker images (#99)
The old workflow for building and testing the code has been adjusted
(and renamed) to also push the Docker image to GitHub's image registry
(ghcr.io) for pushes and manual workflow invocations (but not pull
requests). The images will be tagged "latest" and "sha-$SHA".

If we ever start tagging the code and want to build and tag images with
the Git tags we'll most likely need to make updates to the workflow.

The ARG and LABEL directives in the dockerfile were adjusted somewhat.
Previously we had a "revision" argument that went into both
com.datadoghq.tags.version and org.opencontainers.image.revision.
The OCI label spec stipulates that org.opencontainers.image.revision
should be the VCS identifier, in the Git case the commit SHA-1, but
com.datadoghq.tags.version should probably contain a human-readable
string (e.g. a tag, if available). We've renamed the argument to
"version" and store that value in com.datadoghq.tags.version and
org.opencontainers.image.version. The workflow will pass the Docker
image tag into the "version" argument but also set the
org.opencontainers.image.revision label.

One drawback with the new workflow, that doesn't appear to be trivial
to solve, is that people pushing to their forks in order to create
a PR will trigger a push operation to their ghcr.io space.